% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configHeader}
\alias{configHeader}
\title{set header for configs}
\usage{
configHeader(handle, appIdentifier)
}
\arguments{
\item{handle}{a curl handle used to configure a request}

\item{appIdentifier}{define the application identifier}
}
\description{
set header for configs
}
