% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataHeader}
\alias{dataHeader}
\title{set handle for handle}
\usage{
dataHeader(handle, appIdentifier)
}
\arguments{
\item{handle}{a curl handle used to configure a request}

\item{appIdentifier}{application identifier}
}
\description{
set handle for handle
}
