% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCalls.R
\name{dataTransformJSON2tibble}
\alias{dataTransformJSON2tibble}
\title{transform json into tibble}
\usage{
dataTransformJSON2tibble(resJSON)
}
\arguments{
\item{resJSON}{server response}
}
\description{
transform json into tibble
}
