% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCalls.R
\name{get_dataByParameters}
\alias{get_dataByParameters}
\title{get data by querying parameters}
\usage{
get_dataByParameters(apiPath, appIdentifier, class, loiname, startDate,
  endDate)
}
\arguments{
\item{apiPath}{a curl handle used to configure a request}

\item{appIdentifier}{defines the application identifier}

\item{class}{label to designate the classes}

\item{loiname}{local of interest name}

\item{startDate}{initial date from the interval}

\item{endDate}{final date from the interval}
}
\description{
get data by querying parameters
}
