% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configCalls.R
\name{list_classes}
\alias{list_classes}
\title{list all classes}
\usage{
list_classes(apiPath, appIdentifier)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}
}
\description{
list all classes
}
