% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsCalls.R
\name{list_datasets}
\alias{list_datasets}
\title{set header for configs}
\usage{
list_datasets(apiPath)
}
\arguments{
\item{apiPath}{a curl handle used to configure a request}
}
\description{
set header for configs
}
