% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configCalls.R
\name{list_filters}
\alias{list_filters}
\title{list all filters}
\usage{
list_filters(apiPath, appIdentifier)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}
}
\description{
list all filters
}
