% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configCalls.R
\name{list_local_of_interests}
\alias{list_local_of_interests}
\title{list all local of interest names}
\usage{
list_local_of_interests(apiPath, appIdentifier)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}
}
\description{
list all local of interest names
}
