% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configCalls.R
\name{list_localOfInterestByLocal}
\alias{list_localOfInterestByLocal}
\title{list all loinames by loi}
\usage{
list_localOfInterestByLocal(apiPath, appIdentifier, loi)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}

\item{loi}{define type of local of interest}
}
\description{
list all loinames by loi
}
