% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configCalls.R
\name{list_locals}
\alias{list_locals}
\title{list all locals}
\usage{
list_locals(apiPath, appIdentifier)
}
\arguments{
\item{apiPath}{a curl handle used to configure a request}

\item{appIdentifier}{define the application identifier}
}
\description{
list all locals
}
