% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configCalls.R
\name{list_periods}
\alias{list_periods}
\title{list all periods}
\usage{
list_periods(apiPath, appIdentifier)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}
}
\description{
list all periods
}
