% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_utils.R
\name{tba_appsHeader}
\alias{tba_appsHeader}
\title{set header for apps}
\usage{
tba_appsHeader(handle)
}
\arguments{
\item{handle}{a curl handle used to configure a request}
}
\description{
set header for apps
}
