% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_utils.R
\name{tba_configHeader}
\alias{tba_configHeader}
\title{set header for configs}
\usage{
tba_configHeader(handle, appIdentifier)
}
\arguments{
\item{handle}{a curl handle used to configure a request}

\item{appIdentifier}{define the application identifier}
}
\description{
set header for configs
}
