% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_utils.R
\name{tba_dataHeader}
\alias{tba_dataHeader}
\title{set handle for handle}
\usage{
tba_dataHeader(handle, appIdentifier)
}
\arguments{
\item{handle}{a curl handle used to configure a request}

\item{appIdentifier}{application identifier}
}
\description{
set handle for handle
}
