% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_dataCalls.R
\name{tba_get_dataByLoiname}
\alias{tba_get_dataByLoiname}
\title{list data by loiname}
\usage{
tba_get_dataByLoiname(apiPath, appIdentifier, class, loiname)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}

\item{class}{define the application identifier}

\item{loiname}{define the application identifier}
}
\description{
list data by loiname
}
