% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_dataCalls.R
\name{tba_get_dataByParameters}
\alias{tba_get_dataByParameters}
\title{list data by querying parameters}
\usage{
tba_get_dataByParameters(apiPath, appIdentifier, class, loiname, startDate,
  endDate)
}
\arguments{
\item{apiPath}{a curl handle used to configure a request}

\item{appIdentifier}{defines the application identifier}

\item{class}{label to designate the classes}

\item{loiname}{local of interest name}

\item{startDate}{initial date from the interval}

\item{endDate}{final date from the interval}
}
\description{
list data by querying parameters
}
