% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_appsCalls.R
\name{tba_list_apps_identifier}
\alias{tba_list_apps_identifier}
\title{set header for configs}
\usage{
tba_list_apps_identifier(apiPath, appIdentifier)
}
\arguments{
\item{apiPath}{a curl handle used to configure a request}

\item{appIdentifier}{define the application identifier}
}
\description{
set header for configs
}
