% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_configCalls.R
\name{tba_list_filters}
\alias{tba_list_filters}
\title{list all filters}
\usage{
tba_list_filters(apiPath, appIdentifier)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}
}
\description{
list all filters
}
