% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_configCalls.R
\name{tba_list_loinames}
\alias{tba_list_loinames}
\title{list all loinames}
\usage{
tba_list_loinames(apiPath, appIdentifier)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}
}
\description{
list all loinames
}
