% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_configCalls.R
\name{tba_list_loinamesByloi}
\alias{tba_list_loinamesByloi}
\title{list all loinames by loi}
\usage{
tba_list_loinamesByloi(apiPath, loi, appIdentifier)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{loi}{define type of local of interest}

\item{appIdentifier}{define the application identifier}
}
\description{
list all loinames by loi
}
