% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_configCalls.R
\name{tba_list_lois}
\alias{tba_list_lois}
\title{list all lois}
\usage{
tba_list_lois(apiPath, appIdentifier)
}
\arguments{
\item{apiPath}{a curl handle used to configure a request}

\item{appIdentifier}{define the application identifier}
}
\description{
list all lois
}
