% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tba_request}
\alias{tba_request}
\alias{request}
\title{send request}
\usage{
request(URL, handle)
}
\arguments{
\item{URL}{server URL server}

\item{handle}{a curl handle used to configure a request}
}
\description{
send request
}
