% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCalls.R
\name{get_dataByLocalOfInterest}
\alias{get_dataByLocalOfInterest}
\title{get data by local of interest}
\usage{
get_dataByLocalOfInterest(apiPath, appIdentifier, class, loiname)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}

\item{class}{define the application identifier}

\item{loiname}{define the application identifier}
}
\description{
get data by local of interest
}
