% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tba_configCalls.R
\name{tba_list_loinamesByLoi}
\alias{tba_list_loinamesByLoi}
\title{list all loinames by loi}
\usage{
tba_list_loinamesByLoi(apiPath, appIdentifier, loi)
}
\arguments{
\item{apiPath}{path of terrabrasilis analytics server API}

\item{appIdentifier}{define the application identifier}

\item{loi}{define type of local of interest}
}
\description{
list all loinames by loi
}
